requirements.txt
cryptography
Django
requests
djangorestframework
django-extensions
django-qr-code
django-taggit
django-countries
django-tinymce
django-cors-headers
easy-thumbnails
humanize
mysql
mysql-connector-python
mysqlclient
Pillow
pip
pybase64
sqlparse
traceback2
